requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
faiss-cpu==1.7.4
pypdf==3.17.0
gradio==4.8.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
uvicorn==0.24.0
pydantic==2.5.0

# 新依赖
PyMuPDF==1.23.22
Pillow==10.1.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0

# 可选依赖
streamlit==1.28.2
fastapi==0.104.1
